collabcore-main/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
firebase-admin==6.5.0
python-dotenv==1.0.1
requests==2.31.0
cloudinary>=1.40.0
python-multipart>=0.0.9